% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales_est.R
\name{est_color_pal}
\alias{est_color_pal}
\title{Brand Estonia color palettes}
\usage{
est_color_pal(palette = "group 1", direction = 1, ...)
}
\arguments{
\item{palette}{The name of the color palette.
Use \code{list_color_palettes} to get a list of color palettes.}
}
\description{
Color palettes in the style of Brand Estonia.
}
