% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_est.R
\name{theme_est}
\alias{theme_est}
\title{Brand Estonia theme}
\usage{
theme_est(base_size = 10)
}
\arguments{
\item{base_size}{Base font size.}
}
\description{
A theme inspired by the \href{https://brand.estonia.ee/about/}{Estonian brand platform}.
}
\details{
The Aino typeface is not included in the package because of its terms and conditions.
You can download the typeface from \href{https://toolbox.estonia.ee/media/422}{here}.
In order to use the font,
you need to import it by calling the \code{\link{import_aino_fonts}} function.
}
\examples{
data(funding)
x11() # open graphics device
ggplot(funding, aes(x=factor(year), y=total_raised_capital)) +
  geom_bar(stat="identity") +
  ggtitle("Investments raised by Estonian startups") +
  theme_est()
}
